# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/products": {
      "get": {
        "description": "Get all products",
        "operationId": "ProductController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get all products",
        "tags": [
          "/products"
        ]
      },
      "post": {
        "description": "Create a new product",
        "operationId": "ProductController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Create product",
        "tags": [
          "/products"
        ]
      }
    },
...
    "/products/by-slug/{slug}": {
      "get": {
        "description": "Returns full variant info, parent product, and sibling variants by variant slug.",
        "operationId": "ProductController_getVariantBySlug",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get variant by slug",
        "tags": [
          "/products"
        ]
      }
    },
    "/products/{id}": {
      "get": {
        "description": "Get a single product by id",
        "operationId": "ProductController_findById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get product by id",
        "tags": [
          "/products"
        ]
      },
      "patch": {
        "description": "Update an existing product",
        "operationId": "ProductController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Update product",
        "tags": [
          "/products"
        ]
      },
      "delete": {
        "description": "Delete a product",
        "operationId": "ProductController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Delete product",
        "tags": [
          "/products"
        ]
      }
    },
    "/products/validate": {
      "post": {
        "description": "Check which slugs and SKUs are already taken. Returns arrays of taken values.",
        "operationId": "ProductController_validate",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidateProductDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Validate slugs and SKUs",
        "tags": [
          "/products"
        ]
      }
    },
    "/products/{id}/variants": {
      "get": {
        "description": "Get all variants belonging to a product.",
        "operationId": "ProductController_getVariants",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get product variants",
        "tags": [
          "/products"
        ]
      },
      "post": {
        "description": "Create a new variant for an existing product.",
        "operationId": "ProductController_addVariant",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddVariantDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Add variant to product",
        "tags": [
          "/products"
        ]
      }
    },
    "/products/{id}/variants/{variantId}": {
      "get": {
        "description": "Get a single product variant by id.",
        "operationId": "ProductController_getVariant",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "variantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get variant by id",
        "tags": [
          "/products"
        ]
      },
      "patch": {
        "description": "Partially update a product variant (price, stock, status, etc.).",
        "operationId": "ProductController_updateVariant",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "variantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateVariantDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Update variant",
        "tags": [
          "/products"
        ]
      },
      "delete": {
        "description": "Delete a specific product variant.",
        "operationId": "ProductController_deleteVariant",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "variantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Delete variant",
        "tags": [
          "/products"
        ]
      }
    },
    "/products/{id}/variants/{variantId}/images": {
      "patch": {
        "description": "Replace the images array on a specific product variant.",
        "operationId": "ProductController_setVariantImages",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "variantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetVariantImagesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Set variant images",
        "tags": [
          "/products"
        ]
      }
    },
    "/upload/presign": {
      "post": {
        "description": "Returns presigned S3 PUT URLs and public URLs for direct client-side upload. URLs expire in 15 minutes.",
        "operationId": "UploadController_presign",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PresignUploadDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Generate presigned upload URLs",
        "tags": [
          "/upload"
        ]
      }
    },
    "/upload/confirm": {
      "post": {
        "description": "Verifies that files were successfully uploaded to S3 via headObject checks.",
        "operationId": "UploadController_confirm",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmUploadDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Confirm uploaded files",
        "tags": [
          "/upload"
        ]
      }
    },
    "/upload": {
      "delete": {
        "description": "Permanently deletes one or more files from S3 by their keys.",
        "operationId": "UploadController_delete",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteUploadDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Delete files from S3",
        "tags": [
          "/upload"
        ]
      }
    },
...
  },
  "info": {
    "title": "Urban Tab API",
    "description": "API документація",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "ValidateProductDto": {
        "type": "object",
        "properties": {
          "slugs": {
            "example": [
              "futbolka-bazova-chorna",
              "futbolka-bazova-bila"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "skus": {
            "example": [
              "SKU-001",
              "SKU-002"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "ProductDescriptionDto": {
        "type": "object",
        "properties": {
          "json": {
            "type": "object",
            "example": {
              "ops": [
                {
                  "insert": "Hello"
                }
              ]
            },
            "description": "Quill delta JSON"
          },
          "html": {
            "type": "string",
            "example": "<p>Hello</p>",
            "description": "Rendered HTML"
          }
        },
        "required": [
          "json",
          "html"
        ]
      },
      "VariantTypeDto": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string",
            "example": "color",
            "description": "Variant type key"
          },
          "label": {
            "type": "string",
            "example": "Color",
            "description": "Variant type label"
          }
        },
        "required": [
          "key",
          "label"
        ]
      },
      "AttributeDto": {
        "type": "object",
        "properties": {
          "l": {
            "type": "string",
            "example": "Виробник",
            "description": "Attribute label (human-readable)"
          },
          "v": {
            "type": "object",
            "example": "Sony",
            "description": "Attribute value (string, number, or boolean)"
          }
        },
        "required": [
          "l",
          "v"
        ]
      },
      "CreateVariantDto": {
        "type": "object",
        "properties": {
          "v_value": {
            "type": "object",
            "example": "Червоний",
            "description": "Variant distinguishing value (null for single-variant products)",
            "nullable": true
          },
          "price": {
            "type": "number",
            "example": 440,
            "description": "Price in minor units or base currency"
          },
          "stock": {
            "type": "number",
            "example": 100,
            "description": "Available stock quantity"
          },
          "images": {
            "description": "Image URLs",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "vendor_product_sku": {
            "type": "string",
            "example": "VENDOR-SKU-123",
            "description": "Vendor-assigned SKU"
          }
        },
        "required": [
          "price"
        ]
      },
      "CreateProductDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Wireless Headphones",
            "description": "Product name"
          },
          "category_id": {
            "type": "string",
            "example": "64b1f2c3d4e5f6a7b8c9d0e1",
            "description": "Category ObjectId"
          },
          "subcategory_id": {
            "type": "string",
            "example": "64b1f2c3d4e5f6a7b8c9d0e2",
            "description": "Subcategory ObjectId"
          },
          "vendor_id": {
            "type": "string",
            "example": "64b1f2c3d4e5f6a7b8c9d0e3",
            "description": "Vendor ObjectId"
          },
          "description": {
            "$ref": "#/components/schemas/ProductDescriptionDto"
          },
          "variant_type": {
            "$ref": "#/components/schemas/VariantTypeDto"
          },
          "attributes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AttributeDto"
            }
          },
          "variants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateVariantDto"
            }
          }
        },
        "required": [
          "name",
          "category_id",
          "subcategory_id",
          "vendor_id"
        ]
      },
      "AddVariantDto": {
        "type": "object",
        "properties": {
          "price": {
            "type": "number",
            "example": 440,
            "description": "Variant price"
          },
          "v_value": {
            "type": "object",
            "example": "Чорна",
            "description": "Variant distinguishing value; omit for single-variant products",
            "nullable": true
          },
          "stock": {
            "type": "number",
            "example": 100
          },
          "images": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "vendor_product_sku": {
            "type": "string",
            "example": "VENDOR-SKU-123"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "active",
              "archived"
            ],
            "default": "active"
          }
        },
        "required": [
          "price"
        ]
      },
      "UpdateVariantDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Футболка базова — Чорна"
          },
          "price": {
            "type": "number",
            "example": 440
          },
          "stock": {
            "type": "number",
            "example": 100
          },
          "images": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "v_value": {
            "type": "object",
            "example": "Чорна",
            "nullable": true
          },
          "vendor_product_sku": {
            "type": "string",
            "example": "VENDOR-SKU-123"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "active",
              "archived"
            ]
          }
        }
      },
      "SetVariantImagesDto": {
        "type": "object",
        "properties": {
          "images": {
            "example": [
              "https://cdn.example.com/img.webp"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "images"
        ]
      },
      "UpdateProductDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Wireless Headphones",
            "description": "Product name"
          },
          "category_id": {
            "type": "string",
            "example": "64b1f2c3d4e5f6a7b8c9d0e1"
          },
          "subcategory_id": {
            "type": "string",
            "example": "64b1f2c3d4e5f6a7b8c9d0e2"
          },
          "vendor_id": {
            "type": "string",
            "example": "64b1f2c3d4e5f6a7b8c9d0e3"
          },
          "description": {
            "$ref": "#/components/schemas/ProductDescriptionDto"
          },
          "variant_type": {
            "$ref": "#/components/schemas/VariantTypeDto"
          },
          "attributes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AttributeDto"
            }
          }
        }
      },
      "FilePresignRequestDto": {
        "type": "object",
        "properties": {
          "entityType": {
            "type": "string",
            "example": "product",
            "description": "Entity type the file belongs to (product, user, or vendor)",
            "enum": [
              "product",
              "user",
              "vendor",
              "category"
            ]
          },
          "entityId": {
            "type": "string",
            "example": "507f1f77bcf86cd799439011",
            "description": "MongoDB ObjectId of the entity the file belongs to"
          },
          "contentType": {
            "type": "string",
            "example": "image/webp",
            "description": "MIME type of the file to upload (image/jpeg, image/png, or image/webp)",
            "enum": [
              "image/jpeg",
              "image/png",
              "image/webp"
            ]
          }
        },
        "required": [
          "entityType",
          "entityId",
          "contentType"
        ]
      },
      "PresignUploadDto": {
        "type": "object",
        "properties": {
          "files": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FilePresignRequestDto"
            }
          }
        },
        "required": [
          "files"
        ]
      },
      "ConfirmUploadDto": {
        "type": "object",
        "properties": {
          "keys": {
            "example": "products/abc123/550e8400-e29b-41d4-a716-446655440000.webp",
            "description": "S3 object key identifying the file in the bucket",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "keys"
        ]
      },
      "DeleteUploadDto": {
        "type": "object",
        "properties": {
          "keys": {
            "example": "products/abc123/550e8400-e29b-41d4-a716-446655440000.webp",
            "description": "S3 object key identifying the file in the bucket",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "keys"
        ]
      },
...
    }
  }
}